Replace '<title>TITLE<title>' with subject (render-time)
This is a revision of #280.  It uses a tighter search/replace pattern so
that it only matches "TITLE" when its nested inside `<head>` and `<title>`
elements.  Additionally, it applies the filter at render-time, which means
it will apply to new-or-existing mailings/templates (whether
stock-or-customized; with or without https://github.com/civicrm/mosaico/pull/4).

Note: It is useful to have both:

* With or without this patch, mosaico#4 provides better readability/writeability in the base template.
* With or without mosaico#4, the present commit provides more backward/forward compatibility.

Conflicts:
	mosaico.php